pyproject.toml
[tool.poetry]
name = "fr-code-task"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "fr_code_task"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
django-timezone-field = "^5.0"
drf-spectacular = "^0.24.2"
celery = {extras = ["redis"], version = "^5.2.7"}
django-redis = "^5.2.0"
pytz = "^2022.6"
tzdata = "^2022.6"
django-celery-results = "^2.4.0"
django-celery-beat = "^2.4.0"
redis = "^4.3.4"
psycopg2-binary = "^2.9.5"
requests = "^2.28.1"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
